Fix `Rect::round(_out)` (#157)
"Round these floating-point bounds outwards to integer coordinates" is a
common operation for 2D graphics libraries, and half of the ones I see
implement it incorrectly, as did tiny-skia.

Let's say we have a path that is 2px wide, at x = 0.5:
<img width="429" height="319" alt="image"
src="https://github.com/user-attachments/assets/86d60b9f-21e7-43d0-af75-7b48d038c822"
/>

It at least partially covers columns 0, 1, and 2. That's 3 columns, so
its rounded bounding box is 3 pixels wide. However, taking the ceiling
of the width independent of its position would say it's 2 pixels wide!

We need to take the fractional *position* of the bounding box into
account when calculating its rounded width and height. Or to put it
another way, we need to convert the bottom and right edges into absolute
coordinates before taking the ceiling, and then convert those back into
relative coordinates.

Not changed in this PR, but now I'm suspicious of it: instead of always
rounding the width and height up to 1, should we be returning `None` if
they are not strictly positive?